Retry delete and insert on ambiguity.
Had the logic wrong. When you're on the wrong page, retry on the right
page you must. You must descend the tree again. You must check for
the existence of the current record again. I had it wrong. I checked in
a test that said that wrong was okay, when it obviously was not.

Closes #10.
Closes #9.