Upgrade `advance` to `3.0.0-alpha.8`.
Direction is now a property of iterators. It is one of `mvcc.FORWARD`,
`mvcc.REVERSE` or `mvcc.MAP` defined in the `mvcc` constants module.
Instead of calling `advance.forward()` to construct an interator we
simply call `advance()`. We'd pass a named argument to go in reverse but
we don't go in reverse here.

https://github.com/bigeasy/advance/releases/tag/v3.0.0-alpha.8

Closes #188.